Bug 715113: Update Snappy to r56. r=bent
UltraBlame original commit: b0e65467c4c834dcc2209fed1f9f44bfd157dedb

+++ other-licenses/snappy/src/snappy.h
@@ -144,10 +144,10 @@ namespace snappy {
   
   
   static const int kBlockLog = 15;
-  static const int kBlockSize = 1 << kBlockLog;
+  static const size_t kBlockSize = 1 << kBlockLog;
 
   static const int kMaxHashTableBits = 14;
-  static const int kMaxHashTableSize = 1 << kMaxHashTableBits;
+  static const size_t kMaxHashTableSize = 1 << kMaxHashTableBits;
 
 }
